requirements.txt
asgiref==3.6.0
Django==4.1.7
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fontawesomefree==6.4.0
Pillow==9.5.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
